75: gssi.refactorings -> edelta.refactorings.lib
Task-Url: http://github.com/LorenzoBettini/edelta/issues/75
