+++ edelta.parent/edelta.example/src/AnotherExample.edelta
@@ -1,4 +1,4 @@
-import gssi.refactorings.MMrefactorings
+import edelta.refactorings.lib.MMrefactorings
 
 // IMPORTANT: ecores must be in source directories
 // otherwise you can't refer to them
