+++ edelta.parent/edelta.personlist.example/src/ExampleSyntax1.edelta
@@ -1,4 +1,4 @@
-import gssi.refactorings.MMrefactorings
+import edelta.refactorings.lib.MMrefactorings
 
 // IMPORTANT: ecores must be in a source directory
 // otherwise you can't refer to them
